Revert "Merge builtin-stash-rebase-v2"
This reverts commit d96bb8bc6c636a8869140e860e72e7bdf64bd790, reversing
changes made to fde33e23a8a6e26f6de397609b3d7c360cd777a5.
