+++ .gitignore
@@ -78,9 +78,6 @@
 /git-init-db
 /git-interpret-trailers
 /git-instaweb
-/git-legacy-rebase
-/git-legacy-rebase--interactive
-/git-legacy-stash
 /git-log
 /git-ls-files
 /git-ls-remote
@@ -121,7 +118,7 @@
 /git-read-tree
 /git-rebase
 /git-rebase--am
-/git-rebase--common
+/git-rebase--helper
 /git-rebase--interactive
 /git-rebase--merge
 /git-rebase--preserve-merges
