+++ git-stash.sh
@@ -66,28 +66,6 @@ clear_stash () {
 	fi
 }
 
-maybe_quiet () {
-	case "$1" in
-	--keep-stdout)
-		shift
-		if test -n "$GIT_QUIET"
-		then
-			eval "$@" 2>/dev/null
-		else
-			eval "$@"
-		fi
-		;;
-	*)
-		if test -n "$GIT_QUIET"
-		then
-			eval "$@" >/dev/null 2>&1
-		else
-			eval "$@"
-		fi
-		;;
-	esac
-}
-
 create_stash () {
 	stash_msg=
 	untracked=
@@ -117,18 +95,15 @@ create_stash () {
 	done
 
 	git update-index -q --refresh
-	if maybe_quiet no_changes "$@"
+	if no_changes "$@"
 	then
 		exit 0
 	fi
 
 	# state of the base commit
-	if b_commit=$(maybe_quiet --keep-stdout git rev-parse --verify HEAD)
+	if b_commit=$(git rev-parse --verify HEAD)
 	then
 		head=$(git rev-list --oneline -n 1 HEAD --)
-	elif test -n "$GIT_QUIET"
-	then
-		exit 1
 	else
 		die "$(gettext "You do not have the initial commit yet")"
 	fi
@@ -323,7 +298,7 @@ push_stash () {
 	test -n "$untracked" || git ls-files --error-unmatch -- "$@" >/dev/null || exit 1
 
 	git update-index -q --refresh
-	if maybe_quiet no_changes "$@"
+	if no_changes "$@"
 	then
 		say "$(gettext "No local changes to save")"
 		exit 0
@@ -378,9 +353,6 @@ save_stash () {
 	while test $# != 0
 	do
 		case "$1" in
-		-q|--quiet)
-			GIT_QUIET=t
-			;;
 		--)
 			shift
 			break
